+++ history.md
@@ -1,6 +1,7 @@
 ### History
 
 v2.8.0
+- Added time series record list.
 - Retire hold on options and TS and Settings tabs.
 - Move sync offset button to the map settings panel.
 - Added pending time series selection to the time series selection panel.